// test_utils/parsers/interfaces.ts
/**
 * @license
 * Copyright 2026 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 * @fileoverview Common interfaces used by parsers.
 */

/**
 * Represents a single segment within a stream's Segment Index Box (sidx).
 * This interface is used to store parsed information about each subsegment,
 * including its timing, duration, byte offset within the file, and size.
 */
export interface Segment {
  time: number;
  duration: number;
  offset: number;
  size: number;
}